01-admin-client-localhost/py/requirements.txt
asgiref==3.4.1
cryptocode==0.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
lml==0.1.0
lxml==4.5.1
Markdown==3.3.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
numpy==1.18.4
openpyxl==3.0.7
pandas==1.0.4
protobuf==3.12.2
pycryptodomex==3.10.1
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
pywin32==301
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
xlrd==1.2.0
xlwt==1.3.0
